Move CompareTimingModels to timing.py
